Run the update job a bit more frequently
A complete run takes around 4.5 hours with the current host
configuration, so run it every 6 hours to be safe. Log a message at the
start of the job to see the starting time. Run the cache cleanup job
just before midnight which is the least likely time anything else would
be running.