Specify Python dependency versions explicitly
They are stored in the file requirements-testclutch.txt and correspond
to the actual versions used in development. The testclutch .toml file
specifies tested ranges, but the versions deployed should be exactly
those that being used by developers. The Test Clutch version is not
included in this file, as it may be a prerelease version and is assumed
to match these requirements, regardless of the actual version tag.