Use a virtual environment when building
This avoids touching system packages, although it's really only pip that
falls into this category. All the other packages are installed into a
custom directory (the target tree) so those already leave system
packages alone.  Using a venv is the cleanest way right now to ensure it
all works, but this change could be reverted at such time that pip is
guaranteed to be new enough everywhere this script could run.