Add four more pr comment runs per day
These are timed so that the ingest phase of the incremental update job
is almost always complete, reducing the chance of database writes
happening while jobs are being analyzed. This means a PR comment should
now never take much more than an hour to appear after the CI jobs are
done.